requirements.txt
-i https://pypi.org/simple
absl-py==1.4.0 ; python_version >= '3.6'
anyio==3.6.2 ; python_full_version >= '3.6.2'
astunparse==1.6.3
authlib==1.2.0
cachecontrol==0.12.11 ; python_version >= '3.6'
cachetools==5.3.0 ; python_version ~= '3.7'
camel-converter[pydantic]==3.0.0 ; python_version >= '3.7' and python_version < '4.0'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
cryptography==40.0.2 ; python_version >= '3.6'
decorator==5.1.1 ; python_version >= '3.5'
fastapi==0.95.1
firebase-admin==6.1.0
flatbuffers==23.3.3
gast==0.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core[grpc]==2.11.0 ; platform_python_implementation != 'PyPy'
google-api-python-client==2.86.0 ; python_version >= '3.7'
google-auth==2.17.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0 ; python_version >= '3.6'
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-firestore==2.11.1 ; platform_python_implementation != 'PyPy'
google-cloud-storage==2.8.0 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-pasta==0.2.0
google-resumable-media==2.5.0 ; python_version >= '3.7'
googleapis-common-protos==1.59.0 ; python_version >= '3.7'
grpcio==1.54.0 ; python_version >= '3.7'
grpcio-status==1.54.0
h11==0.14.0 ; python_version >= '3.7'
h5py==3.8.0 ; python_version >= '3.7'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4 ; python_version >= '3.5'
jax==0.4.8 ; python_version >= '3.8'
keras==2.12.0 ; python_version >= '3.8'
libclang==16.0.0
markdown==3.4.3 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
meilisearch==0.26.0
ml-dtypes==0.1.0 ; python_version >= '3.7'
msgpack==1.0.5
numpy==1.23.5 ; python_version >= '3.8'
oauthlib==3.2.2 ; python_version >= '3.6'
opt-einsum==3.3.0 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
pillow==9.5.0
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.22.3 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==1.10.7 ; python_version >= '3.7'
pyjwt[crypto]==2.6.0 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_version >= '3.1'
python-dotenv==1.0.0
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
scipy==1.10.1 ; python_version < '3.12' and python_version >= '3.8'
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.26.1 ; python_version >= '3.7'
tensorboard==2.12.2 ; python_version >= '3.8'
tensorboard-data-server==0.7.0 ; python_version >= '3.7'
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0 ; python_version >= '3.7'
tensorflow-intel==2.12.0 ; platform_system == 'Windows'
tensorflow-io-gcs-filesystem==0.31.0 ; platform_machine != 'arm64' or platform_system != 'Darwin'
termcolor==2.3.0 ; python_version >= '3.7'
tqdm==4.65.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.22.0
validators==0.20.0 ; python_version >= '3.4'
weaviate-client==3.16.2
werkzeug==2.3.1 ; python_version >= '3.8'
wheel==0.40.0 ; python_version >= '3.7'
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
